project-openweather/consumer/requirements.txt
kafka-python
pymongo
python-dotenv
